+++ generic/query_generated.go
@@ -137,6 +137,14 @@ func (q *Query0) Relation() ecs.Entity {
 	return q.Query.Relation(ecs.ID(q.target))
 }
 
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query0.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query0) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
+}
+
 //////////////////////////////////////////////////////////////////////////
 
 // Filter1 is a helper for building [Query1] query iterators.
@@ -292,6 +300,14 @@ func (q *Query1[A]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query1.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query1[A]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -453,6 +469,14 @@ func (q *Query2[A, B]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query2.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query2[A, B]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -618,6 +642,14 @@ func (q *Query3[A, B, C]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query3.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query3[A, B, C]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -787,6 +819,14 @@ func (q *Query4[A, B, C, D]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query4.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query4[A, B, C, D]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -960,6 +1000,14 @@ func (q *Query5[A, B, C, D, E]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query5.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query5[A, B, C, D, E]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -1137,6 +1185,14 @@ func (q *Query6[A, B, C, D, E, F]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query6.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query6[A, B, C, D, E, F]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -1318,6 +1374,14 @@ func (q *Query7[A, B, C, D, E, F, G]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query7.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query7[A, B, C, D, E, F, G]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -1503,4 +1567,12 @@ func (q *Query8[A, B, C, D, E, F, G, H]) Relation() ecs.Entity {
 		panic("query has no relation")
 	}
 	return q.Query.Relation(ecs.ID(q.target))
+}
+
+// RelationUnchecked returns the target entity for the query's relation.
+//
+// RelationUnchecked is an optimized version of [Query8.Relation].
+// Does not check that the component ID is applicable.
+func (q *Query8[A, B, C, D, E, F, G, H]) RelationUnchecked() ecs.Entity {
+	return q.Query.RelationUnchecked(ecs.ID(q.target))
 }