+++ generic/query_test.go
@@ -123,7 +123,7 @@ func TestQuery0(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter0().
@@ -135,6 +135,7 @@ func TestQuery0(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -192,7 +193,7 @@ func TestQuery1(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter1[testRelationA]().
@@ -203,6 +204,7 @@ func TestQuery1(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -267,7 +269,7 @@ func TestQuery2(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter2[testStruct0, testRelationA]().
@@ -278,6 +280,7 @@ func TestQuery2(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -345,7 +348,7 @@ func TestQuery3(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter3[testStruct0, testStruct1, testRelationA]().
@@ -356,6 +359,7 @@ func TestQuery3(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -428,7 +432,7 @@ func TestQuery4(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter4[
@@ -441,6 +445,7 @@ func TestQuery4(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -518,7 +523,7 @@ func TestQuery5(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter5[
@@ -532,6 +537,7 @@ func TestQuery5(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -614,7 +620,7 @@ func TestQuery6(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter6[
@@ -628,6 +634,7 @@ func TestQuery6(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -718,7 +725,7 @@ func TestQuery7(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter7[
@@ -732,6 +739,7 @@ func TestQuery7(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -819,7 +827,7 @@ func TestQuery8(t *testing.T) {
 	w.Add(e1, relID)
 	w.Add(e2, relID)
 
-	w.SetRelation(e0, relID, targ)
+	w.Relations().Set(e0, relID, targ)
 
 	filter2 :=
 		NewFilter8[
@@ -833,6 +841,7 @@ func TestQuery8(t *testing.T) {
 	for q.Next() {
 		trg := q.Relation()
 		assert.Equal(t, targ, trg)
+		assert.Equal(t, targ, q.RelationUnchecked())
 	}
 
 	filter2.Register(&w)
@@ -873,7 +882,7 @@ func TestQueryGeneric(t *testing.T) {
 
 	target := world.NewEntity(rotID)
 
-	world.SetRelation(entities[0], relID, target)
+	world.Relations().Set(entities[0], relID, target)
 
 	filter := NewFilter2[testStruct2, testRelationA]().WithRelation(T[testRelationA](), target)
 	q2 := filter.Query(&world)